llm-manager/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
elasticsearch==8.11.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
httpx==0.25.2
asyncio-mqtt==0.16.1 